New Collection button fix
Fixed issue on the new collection button were it would not work if the user navigated away from the collection view and then back to the collection view. This was caused by the two cancellable variables not being initiated when the images are loaded from CoreData.

Various spaces and breakpoints removed.

Random page variable was moved as the same set of images was shown everytime the New collection button was pressed